requirements.txt
crewai>=1.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
streamlit>=1.32.0